Added all DTO for requests

+++ WebShop.DAL1/Entities/Products.cs
@@ -3,13 +3,10 @@ namespace WebShop.DAL.Entities
 {
     public class Products
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string? Name { get; set; }
-
         public Categories? Categories { get; set; } 
-
         public string? Description { get; set; }
-
         public double? Price { get; set; }
     }
 }
